requirements_face_api.txt
flask
flask-cors
torch
torchvision
ultralytics
opencv-python<4.10
pillow
numpy<2
